# Use local RNG in SampleReducer