@desa 2016-12-14
@mark-rushakoff any reason why the appveyor tests are taking so long?