@mark-rushakoff 2016-12-14
Looks like something wrong on AppVeyor's side. I'll force push this branch in hopes of re-triggering a build.