@desa 2016-12-14
LGTM. Pending Appveyor actually working.