@mark-rushakoff 2016-12-14
@desa the AllSamplesSeen test is failing repeatedly on Appveyor, even though I bumped up the iterations from 6 to now 24. It seems like increasing it more is masking an underlying issue. Thoughts?