@desa 2016-12-15
Very weird. Is there a way to ssh into the container like on Circle? If we change it back to using the rng with a mutex does the problem still exist?