@mark-rushakoff 2016-12-15
I don't think you can debug the containers on Appveyor. 

I think it was passing before because:
* Go tests run in their own executable per package
* The test was using the global RNG, so it was working off the same seed every time, which happened to pass the test at 6 iterations

I could believe an occasional flaky run at 6 or 24 iterations due to "bad luck" with the RNG, but it's very concerning that it's failed multiple times in a row at 24 iterations with what appears to be a properly seeded, local RNG.

I can try a separate branch with the global RNG again, perhaps seeding it, but I'd expect that to behave the same as it does currently.