@desa 2016-12-15
Yeah I completely agree. The thing that I find odd is that it isn't an issue with the Circle build. Maybe it's a windows thing?